moa: orchestrate from any node, build worker pool from mesh-wide gossip
Before this change MoA only ran when the request hit `api_proxy` on a
serving host. A pure `--client` node received `model: "mesh"` in
`handle_mesh_request`, fell through to the "forward to any host"
fallback, and the receiving host either ran an older binary that
ignored the "mesh" name or built a single-model config because its
local `ModelTargets` only had its own model. End result: no fan-out
happened, MoA was effectively dead from any client node, and the
worker pool depended on which node received the request rather than
on what was actually in the mesh.

Two structural fixes:

1. New `moa_gateway` module owns the intercept. Both `api_proxy` and
   `handle_mesh_request` now call `try_handle_moa` — the request is
   handled wherever it lands, whether the node serves models locally
   or not.

2. `build_moa_config` enumerates `Node::models_being_served()` (the
   mesh-wide union of local + gossip) instead of the local routing
   table. Locally-served models are wired directly to the skippy port
   via the routing table when one is present (host mode); everything
   else opens a QUIC tunnel to the hash-preferred peer that advertises
   the model. On a pure client every worker is remote.

Side effects:
- Canonical-base dedup now strips an `@branch` segment without losing
  the trailing quant tag, so `unsloth/Qwen3-8B-GGUF@main:Q4_K_M` and
  `Qwen3-8B-Q4_K_M` collapse to one worker instead of being treated
  as two distinct models.
- The MoA-related backends (`LocalModelBackend`, `RemoteModelBackend`)
  and the SSE wrapper moved out of `ingress.rs` into the shared
  module; `ingress.rs` shrinks by ~400 lines.

Verified live against the public mesh from a `--client --auto` node:
- Chat completion: `x-moa-workers: 4` (all 4 mesh-wide models),
  early-exit path, correct answer.
- Tool-equipped request: 3 tool-capable workers, early-exit path,
  correct shape.
