+++ crates/mesh-llm-host-runtime/src/network/openai/ingress.rs
@@ -257,76 +257,18 @@ pub(crate) async fn api_proxy(
                     }
 
                     // ── MoA intercept ─────────────────────────────────
-                    // When model == "mesh", fan out to all available models
-                    // via the MoA gateway.  Local models are called directly
-                    // on their skippy port.  Remote models go through the
-                    // QUIC tunnel.
+                    // When model == "mesh", fan out to all mesh-wide
+                    // available models via the MoA gateway. The local
+                    // routing table tells us which models can be reached
+                    // directly on a skippy port; everything else goes over
+                    // QUIC.
                     if effective_model.as_deref() == Some(moa::VIRTUAL_MODEL_NAME) {
-                        request.ensure_body_json();
-                        if let Some(body_json) = request.body_json.clone() {
-                            if let Some(config) = build_moa_config(&targets, &node, &callable).await
-                            {
-                                let was_streaming = body_json
-                                    .get("stream")
-                                    .and_then(|v| v.as_bool())
-                                    .unwrap_or(false);
-                                let mut moa_body = body_json;
-                                moa_body.as_object_mut().map(|o| o.remove("stream"));
-
-                                let moa_result = moa::handle_turn(&config, &moa_body).await;
-
-                                let workers_ok = moa_result
-                                    .worker_summaries
-                                    .iter()
-                                    .filter(|w| w.succeeded)
-                                    .count();
-                                let workers_total = moa_result.worker_summaries.len();
-                                tracing::info!(
-                                    "moa: {}ms, {}/{} workers, kind={}, reducer={} (attempts={})",
-                                    moa_result.elapsed_ms,
-                                    workers_ok,
-                                    workers_total,
-                                    moa_result.turn_kind.label(),
-                                    moa_result.reducer_used,
-                                    moa_result.reducer_attempts,
-                                );
-
-                                // Observability headers — benches and ops
-                                // tooling can read these without parsing the
-                                // body. Ignored by normal OpenAI clients.
-                                let extra_headers: Vec<(&str, String)> = vec![
-                                    ("x-moa-elapsed-ms", moa_result.elapsed_ms.to_string()),
-                                    ("x-moa-turn", moa_result.turn_kind.label().to_string()),
-                                    ("x-moa-workers", workers_total.to_string()),
-                                    ("x-moa-workers-ok", workers_ok.to_string()),
-                                    ("x-moa-reducer", moa_result.reducer_used.to_string()),
-                                    (
-                                        "x-moa-reducer-attempts",
-                                        moa_result.reducer_attempts.to_string(),
-                                    ),
-                                ];
-
-                                if was_streaming {
-                                    let _ = send_moa_as_sse(
-                                        tcp_stream,
-                                        &moa_result.response_body,
-                                        &extra_headers,
-                                    )
-                                    .await;
-                                } else {
-                                    let _ = proxy::send_json_ok_with_headers(
-                                        tcp_stream,
-                                        &moa_result.response_body,
-                                        &extra_headers,
-                                    )
-                                    .await;
-                                }
-                                return;
-                            }
-                        }
-                        let _ = proxy::send_503(
+                        let _ = crate::network::openai::moa_gateway::try_handle_moa(
+                            &node,
                             tcp_stream,
-                            "MoA requires ≥2 models available in the mesh",
+                            &mut request,
+                            effective_model.as_deref(),
+                            Some(&targets),
                         )
                         .await;
                         return;
@@ -599,375 +541,6 @@ fn has_available_candidates(targets: &election::ModelTargets, model: &str) -> bo
         .candidates(model)
         .iter()
         .any(|target| !matches!(target, election::InferenceTarget::None))
-}
-
-/// Wrap a completed MoA chat-completion response as SSE frames so streaming
-/// clients (like Goose) can consume it.  Emits one delta chunk with the full
-/// content, then a `finish_reason: stop` chunk, then `[DONE]`.
-///
-/// `extra_headers` are emitted alongside the standard SSE response headers
-/// (used to attach `x-moa-*` observability headers).
-async fn send_moa_as_sse(
-    mut stream: tokio::net::TcpStream,
-    response: &serde_json::Value,
-    extra_headers: &[(&str, String)],
-) -> std::io::Result<()> {
-    use tokio::io::AsyncWriteExt;
-
-    let mut header = String::from(
-        "HTTP/1.1 200 OK\r\nContent-Type: text/event-stream\r\nTransfer-Encoding: chunked\r\nCache-Control: no-cache\r\nConnection: close\r\n",
-    );
-    for (name, value) in extra_headers {
-        header.push_str(name);
-        header.push_str(": ");
-        header.push_str(value);
-        header.push_str("\r\n");
-    }
-    header.push_str("\r\n");
-    stream.write_all(header.as_bytes()).await?;
-
-    let id = response
-        .get("id")
-        .and_then(|v| v.as_str())
-        .unwrap_or("chatcmpl-mesh");
-    let model = response
-        .get("model")
-        .and_then(|v| v.as_str())
-        .unwrap_or(moa::VIRTUAL_MODEL_NAME);
-    let raw_content = response
-        .pointer("/choices/0/message/content")
-        .and_then(|v| v.as_str())
-        .unwrap_or("");
-    // Strip <think>...</think> tags so streaming clients get clean content
-    let content = strip_think_from_content(raw_content);
-
-    // Check if this is a tool_calls response
-    let tool_calls = response
-        .pointer("/choices/0/message/tool_calls")
-        .and_then(|v| v.as_array())
-        .cloned();
-
-    let finish_reason = if tool_calls.is_some() {
-        "tool_calls"
-    } else {
-        "stop"
-    };
-
-    // Content or tool_calls chunk
-    let delta = if let Some(ref tcs) = tool_calls {
-        // For tool_calls, emit them in the delta
-        serde_json::json!({
-            "role": "assistant",
-            "tool_calls": tcs.iter().enumerate().map(|(i, tc)| {
-                serde_json::json!({
-                    "index": i,
-                    "id": tc.get("id").and_then(|v| v.as_str()).unwrap_or("call_0"),
-                    "type": "function",
-                    "function": tc.get("function").cloned().unwrap_or(serde_json::json!({})),
-                })
-            }).collect::<Vec<_>>()
-        })
-    } else {
-        serde_json::json!({ "role": "assistant", "content": content })
-    };
-
-    let chunk = serde_json::json!({
-        "id": id,
-        "object": "chat.completion.chunk",
-        "model": model,
-        "choices": [{
-            "index": 0,
-            "delta": delta,
-            "finish_reason": null,
-        }]
-    });
-    let data = format!("data: {}\n\n", chunk);
-    let framed = format!("{:x}\r\n{}\r\n", data.len(), data);
-    stream.write_all(framed.as_bytes()).await?;
-
-    // Stop chunk
-    let stop = serde_json::json!({
-        "id": id,
-        "object": "chat.completion.chunk",
-        "model": model,
-        "choices": [{
-            "index": 0,
-            "delta": {},
-            "finish_reason": finish_reason,
-        }]
-    });
-    let data = format!("data: {}\n\n", stop);
-    let framed = format!("{:x}\r\n{}\r\n", data.len(), data);
-    stream.write_all(framed.as_bytes()).await?;
-
-    // Done
-    let done = "data: [DONE]\n\n";
-    let framed = format!("{:x}\r\n{}\r\n", done.len(), done);
-    stream.write_all(framed.as_bytes()).await?;
-
-    // Terminating chunk
-    stream.write_all(b"0\r\n\r\n").await?;
-    stream.shutdown().await?;
-    Ok(())
-}
-
-/// Strip `<think>...</think>` tags and orphan `</think>` from content.
-fn strip_think_from_content(text: &str) -> String {
-    let mut result = text.to_string();
-    while let Some(start) = result.find("<think>") {
-        if let Some(end) = result[start..].find("</think>") {
-            result = format!(
-                "{}{}",
-                &result[..start],
-                &result[start + end + "</think>".len()..]
-            );
-        } else {
-            result = result[..start].to_string();
-            break;
-        }
-    }
-    result = result.replace("</think>", "");
-    result.trim().to_string()
-}
-
-/// Backend that calls a local model directly on its skippy HTTP port.
-struct LocalModelBackend {
-    port: u16,
-    http: reqwest::Client,
-}
-
-#[async_trait::async_trait]
-impl moa::ModelBackend for LocalModelBackend {
-    async fn chat_completion(
-        &self,
-        model: &str,
-        messages: &[serde_json::Value],
-        tools: Option<&serde_json::Value>,
-        max_tokens: u32,
-        timeout: std::time::Duration,
-        sampling: moa::SamplingParams,
-    ) -> Result<serde_json::Value, String> {
-        let url = format!("http://127.0.0.1:{}/v1/chat/completions", self.port);
-        let mut body = serde_json::json!({
-            "model": model,
-            "messages": messages,
-            "max_tokens": max_tokens,
-            "temperature": sampling.temperature,
-            "top_p": sampling.top_p,
-            "stream": false,
-            "mesh_hooks": false,
-        });
-        if let Some(tools) = tools {
-            body.as_object_mut()
-                .unwrap()
-                .insert("tools".to_string(), tools.clone());
-        }
-        let resp = self
-            .http
-            .post(&url)
-            .json(&body)
-            .timeout(timeout)
-            .send()
-            .await
-            .map_err(|e| format!("local:{} failed: {e}", self.port))?;
-        let status = resp.status();
-        if !status.is_success() {
-            let text = resp.text().await.unwrap_or_default();
-            return Err(format!("HTTP {status}: {}", &text[..text.len().min(200)]));
-        }
-        resp.json::<serde_json::Value>()
-            .await
-            .map_err(|e| format!("parse: {e}"))
-    }
-}
-
-/// Backend that calls a remote model over the QUIC tunnel.
-struct RemoteModelBackend {
-    node: mesh::Node,
-    peer_id: iroh::EndpointId,
-}
-
-#[async_trait::async_trait]
-impl moa::ModelBackend for RemoteModelBackend {
-    async fn chat_completion(
-        &self,
-        model: &str,
-        messages: &[serde_json::Value],
-        tools: Option<&serde_json::Value>,
-        max_tokens: u32,
-        timeout: std::time::Duration,
-        sampling: moa::SamplingParams,
-    ) -> Result<serde_json::Value, String> {
-        let mut body = serde_json::json!({
-            "model": model,
-            "messages": messages,
-            "max_tokens": max_tokens,
-            "temperature": sampling.temperature,
-            "top_p": sampling.top_p,
-            "stream": false,
-            "mesh_hooks": false,
-        });
-        if let Some(tools) = tools {
-            body.as_object_mut()
-                .unwrap()
-                .insert("tools".to_string(), tools.clone());
-        }
-        let body_bytes = serde_json::to_vec(&body).map_err(|e| format!("serialize: {e}"))?;
-        let http_request = format!(
-            "POST /v1/chat/completions HTTP/1.1\r\n\
-             Host: localhost\r\n\
-             Content-Type: application/json\r\n\
-             Content-Length: {}\r\n\
-             \r\n",
-            body_bytes.len()
-        );
-        let mut raw = http_request.into_bytes();
-        raw.extend_from_slice(&body_bytes);
-
-        let result = tokio::time::timeout(timeout, async {
-            let (mut send, mut recv) = self
-                .node
-                .open_http_tunnel(self.peer_id)
-                .await
-                .map_err(|e| format!("tunnel: {e}"))?;
-            send.write_all(&raw)
-                .await
-                .map_err(|e| format!("send: {e}"))?;
-            send.finish().map_err(|e| format!("finish: {e}"))?;
-            let response = recv
-                .read_to_end(256 * 1024)
-                .await
-                .map_err(|e| format!("recv: {e}"))?;
-            parse_quic_http_response(&response)
-        })
-        .await
-        .map_err(|_| format!("remote timeout after {}s", timeout.as_secs()))?;
-        result
-    }
-}
-
-fn parse_quic_http_response(response: &[u8]) -> Result<serde_json::Value, String> {
-    let s = String::from_utf8_lossy(response);
-    let header_end = s
-        .find("\r\n\r\n")
-        .ok_or_else(|| "malformed HTTP response".to_string())?;
-    let status_line = s[..header_end].lines().next().unwrap_or("");
-    let status: u16 = status_line
-        .split_whitespace()
-        .nth(1)
-        .and_then(|s| s.parse().ok())
-        .unwrap_or(0);
-    if status != 200 {
-        return Err(format!("HTTP {status}: {}", &s[..s.len().min(200)]));
-    }
-    let body = &s[header_end + 4..];
-    serde_json::from_str(body).map_err(|e| format!("parse: {e}"))
-}
-
-/// Build MoA config with mesh-native backends.  Returns None if <2 models.
-async fn build_moa_config(
-    targets: &election::ModelTargets,
-    node: &mesh::Node,
-    callable: &[String],
-) -> Option<moa::GatewayConfig> {
-    let http = reqwest::Client::new();
-    let mut seen_bases = std::collections::HashSet::new();
-    let mut backends: Vec<std::sync::Arc<dyn moa::ModelBackend>> = Vec::new();
-    let mut models: Vec<moa::ModelEntry> = Vec::new();
-
-    // Sort by name length so shorter (canonical) names win dedup
-    let mut sorted: Vec<&String> = callable.iter().collect();
-    sorted.sort_by_key(|n| n.len());
-
-    for name in sorted {
-        if name == moa::VIRTUAL_MODEL_NAME {
-            continue;
-        }
-        let base = name
-            .split('@')
-            .next()
-            .unwrap_or(name)
-            .to_lowercase()
-            .replace("-gguf", "")
-            .replace("unsloth/", "")
-            .replace("meshllm/", "");
-        if !seen_bases.insert(base) {
-            continue;
-        }
-
-        // Try local target first
-        let local_port = targets.targets.get(name.as_str()).and_then(|tv| {
-            tv.iter().find_map(|t| match t {
-                election::InferenceTarget::Local(p) => Some(*p),
-                _ => None,
-            })
-        });
-
-        if let Some(port) = local_port {
-            let backend_idx = backends.len();
-            backends.push(std::sync::Arc::new(LocalModelBackend {
-                port,
-                http: http.clone(),
-            }));
-            models.push(moa::ModelEntry {
-                name: name.clone(),
-                backend_index: backend_idx,
-            });
-            continue;
-        }
-
-        // Try remote
-        let remote_peer = targets.targets.get(name.as_str()).and_then(|tv| {
-            tv.iter().find_map(|t| match t {
-                election::InferenceTarget::Remote(id) => Some(*id),
-                _ => None,
-            })
-        });
-
-        if let Some(peer_id) = remote_peer {
-            let backend_idx = backends.len();
-            backends.push(std::sync::Arc::new(RemoteModelBackend {
-                node: node.clone(),
-                peer_id,
-            }));
-            models.push(moa::ModelEntry {
-                name: name.clone(),
-                backend_index: backend_idx,
-            });
-            continue;
-        }
-
-        // Also check hosts_for_model for models not yet in the routing table
-        let remote_hosts = node.hosts_for_model(name).await;
-        if let Some(peer_id) = remote_hosts.into_iter().next() {
-            let backend_idx = backends.len();
-            backends.push(std::sync::Arc::new(RemoteModelBackend {
-                node: node.clone(),
-                peer_id,
-            }));
-            models.push(moa::ModelEntry {
-                name: name.clone(),
-                backend_index: backend_idx,
-            });
-        }
-    }
-
-    if models.len() < 2 {
-        return None;
-    }
-
-    Some(moa::GatewayConfig {
-        backends,
-        models,
-        worker_timeout: std::time::Duration::from_secs(15),
-        // Per-attempt cap; hedged_reducer_call hedges across candidates so the
-        // end-to-end wait is roughly reducer_timeout + a couple of hedge delays.
-        reducer_timeout: std::time::Duration::from_secs(15),
-        // Start a second reducer candidate after 5s if the first hasn't replied
-        // (or sooner on outright failure). Cheap on the happy path, big win on
-        // the cold-KV / stale-peer tail.
-        hedge_delay: std::time::Duration::from_secs(5),
-    })
 }
 
 pub(crate) fn callable_models(targets: &election::ModelTargets) -> Vec<String> {
