+++ crates/mesh-llm-host-runtime/src/network/openai/moa_gateway.rs
@@ -1,0 +1,543 @@
+//! Mesh-wide MoA orchestration entrypoint.
+//!
+//! Any node that receives a chat-completion request with `model: "mesh"`
+//! runs MoA orchestration here, regardless of whether that node is serving
+//! models locally. The worker pool is built from gossip — every model
+//! advertised by any peer (or hosted locally) is a candidate.
+//!
+//! Both the host's `api_proxy` and the passive `handle_mesh_request` path
+//! call `try_handle_moa`. On a pure client node, all backends are remote;
+//! on a serving host, the local model is wired directly to its skippy port
+//! and the rest go over QUIC.
+
+use crate::inference::election;
+use crate::mesh;
+use crate::network::openai::transport as proxy;
+use mesh_mixture_of_agents as moa;
+use tokio::io::AsyncWriteExt;
+use tokio::net::TcpStream;
+
+/// Detect `model: "mesh"`, build a mesh-wide MoA config, run the turn,
+/// send the HTTP response (JSON or SSE), and return `true` if the request
+/// was handled. Returns `false` if the request is not for MoA, so the
+/// caller can fall through to normal routing.
+///
+/// On MoA failure (e.g. <2 models in the mesh) sends a 503 and still
+/// returns `true` — the caller must not also try to respond.
+pub async fn try_handle_moa(
+    node: &mesh::Node,
+    tcp_stream: TcpStream,
+    request: &mut proxy::BufferedHttpRequest,
+    effective_model: Option<&str>,
+    targets: Option<&election::ModelTargets>,
+) -> Option<TcpStream> {
+    if effective_model != Some(moa::VIRTUAL_MODEL_NAME) {
+        return Some(tcp_stream);
+    }
+
+    request.ensure_body_json();
+    let Some(body_json) = request.body_json.clone() else {
+        let _ = proxy::send_400(tcp_stream, "MoA requires a JSON body").await;
+        return None;
+    };
+
+    let Some(config) = build_moa_config(node, targets).await else {
+        let _ = proxy::send_503(tcp_stream, "MoA requires ≥2 models available in the mesh").await;
+        return None;
+    };
+
+    let was_streaming = body_json
+        .get("stream")
+        .and_then(|v| v.as_bool())
+        .unwrap_or(false);
+    let mut moa_body = body_json;
+    moa_body.as_object_mut().map(|o| o.remove("stream"));
+
+    let moa_result = moa::handle_turn(&config, &moa_body).await;
+
+    let workers_ok = moa_result
+        .worker_summaries
+        .iter()
+        .filter(|w| w.succeeded)
+        .count();
+    let workers_total = moa_result.worker_summaries.len();
+    tracing::info!(
+        "moa: {}ms, {}/{} workers, kind={}, reducer={} (attempts={})",
+        moa_result.elapsed_ms,
+        workers_ok,
+        workers_total,
+        moa_result.turn_kind.label(),
+        moa_result.reducer_used,
+        moa_result.reducer_attempts,
+    );
+
+    let extra_headers: Vec<(&str, String)> = vec![
+        ("x-moa-elapsed-ms", moa_result.elapsed_ms.to_string()),
+        ("x-moa-turn", moa_result.turn_kind.label().to_string()),
+        ("x-moa-workers", workers_total.to_string()),
+        ("x-moa-workers-ok", workers_ok.to_string()),
+        ("x-moa-reducer", moa_result.reducer_used.to_string()),
+        (
+            "x-moa-reducer-attempts",
+            moa_result.reducer_attempts.to_string(),
+        ),
+    ];
+
+    if was_streaming {
+        let _ = send_moa_as_sse(tcp_stream, &moa_result.response_body, &extra_headers).await;
+    } else {
+        let _ =
+            proxy::send_json_ok_with_headers(tcp_stream, &moa_result.response_body, &extra_headers)
+                .await;
+    }
+    None
+}
+
+/// Build a MoA gateway config from this node's mesh-wide view.
+///
+/// Every distinct model in the mesh becomes a worker:
+/// - Models served by this node → `LocalModelBackend` (direct skippy port)
+/// - Models served by a peer → `RemoteModelBackend` (QUIC tunnel)
+///
+/// Models are deduplicated by canonical base name so e.g.
+/// `unsloth/Qwen3-8B-GGUF:Q4_K_M` and `Qwen3-8B-Q4_K_M` (different naming
+/// conventions for the same model from different peers) only show up once.
+///
+/// Returns `None` if fewer than 2 distinct models exist — MoA needs at
+/// least two workers to be meaningfully different from a single call.
+///
+/// `targets` is the runtime's local routing table, used to discover the
+/// skippy port for locally-served models. In passive (`--client`) mode
+/// this is `None` — every backend goes over QUIC. In `serve` mode it's
+/// `Some`, so locally-served models bypass the tunnel.
+pub async fn build_moa_config(
+    node: &mesh::Node,
+    targets: Option<&election::ModelTargets>,
+) -> Option<moa::GatewayConfig> {
+    let http = reqwest::Client::new();
+    let mut seen_bases = std::collections::HashSet::new();
+    let mut backends: Vec<std::sync::Arc<dyn moa::ModelBackend>> = Vec::new();
+    let mut models: Vec<moa::ModelEntry> = Vec::new();
+    let mut local_count = 0usize;
+
+    // Full mesh-wide model list (local + every peer's advertised
+    // routable models). Sorted by name length so the shorter (canonical)
+    // form wins dedup.
+    let mut all_models: Vec<String> = node.models_being_served().await;
+    all_models.sort_by_key(|n| n.len());
+
+    for name in all_models {
+        if name == moa::VIRTUAL_MODEL_NAME {
+            continue;
+        }
+        let base = canonical_base_name(&name);
+        if !seen_bases.insert(base) {
+            continue;
+        }
+
+        // Prefer local skippy port when this node serves the model.
+        let local_port = targets.and_then(|t| {
+            t.targets.get(name.as_str()).and_then(|tv| {
+                tv.iter().find_map(|t| match t {
+                    election::InferenceTarget::Local(p) => Some(*p),
+                    _ => None,
+                })
+            })
+        });
+        if let Some(port) = local_port {
+            let backend_idx = backends.len();
+            backends.push(std::sync::Arc::new(LocalModelBackend {
+                port,
+                http: http.clone(),
+            }));
+            models.push(moa::ModelEntry {
+                name: name.clone(),
+                backend_index: backend_idx,
+            });
+            local_count += 1;
+            continue;
+        }
+
+        // Otherwise find a remote host. hosts_for_model returns peers in
+        // hash-preferred order; take the first.
+        let remote_hosts = node.hosts_for_model(&name).await;
+        if let Some(peer_id) = remote_hosts.into_iter().next() {
+            let backend_idx = backends.len();
+            backends.push(std::sync::Arc::new(RemoteModelBackend {
+                node: node.clone(),
+                peer_id,
+            }));
+            models.push(moa::ModelEntry {
+                name: name.clone(),
+                backend_index: backend_idx,
+            });
+        }
+    }
+
+    if models.len() < 2 {
+        tracing::warn!(
+            "MoA: only {} model(s) reachable, need ≥2 (models={:?})",
+            models.len(),
+            models.iter().map(|m| &m.name).collect::<Vec<_>>()
+        );
+        return None;
+    }
+
+    tracing::info!(
+        "MoA config: {} workers ({} local, {} remote): {:?}",
+        models.len(),
+        local_count,
+        models.len() - local_count,
+        models.iter().map(|m| m.name.as_str()).collect::<Vec<_>>(),
+    );
+
+    Some(moa::GatewayConfig {
+        backends,
+        models,
+        worker_timeout: std::time::Duration::from_secs(15),
+        // Per-attempt cap; hedged_reducer_call hedges across candidates so the
+        // end-to-end wait is roughly reducer_timeout + a couple of hedge delays.
+        reducer_timeout: std::time::Duration::from_secs(15),
+        // Start a second reducer candidate after 5s if the first hasn't replied
+        // (or sooner on outright failure). Cheap on the happy path, big win on
+        // the cold-KV / stale-peer tail.
+        hedge_delay: std::time::Duration::from_secs(5),
+    })
+}
+
+/// Canonical name used for cross-peer dedup. Different peers advertise the
+/// same model under different conventions (`unsloth/Qwen3-8B-GGUF:Q4_K_M`
+/// vs `Qwen3-8B-Q4_K_M`); normalize before comparing.
+///
+/// Strategy: strip the publisher prefix, the `-gguf` suffix, any `@branch`
+/// suffix, then keep only `[a-z0-9]` characters so `:` vs `-` separators
+/// don't matter.
+fn canonical_base_name(name: &str) -> String {
+    let lower = name.to_lowercase();
+    // Drop an `@branch` segment if present, keeping anything after the
+    // next `:` so quant tags survive (e.g. `repo@main:q4_k_m` → `repo:q4_k_m`).
+    let no_branch = match lower.find('@') {
+        Some(at) => {
+            let after = &lower[at + 1..];
+            let rest = after.find(':').map(|c| &after[c..]).unwrap_or("");
+            format!("{}{}", &lower[..at], rest)
+        }
+        None => lower,
+    };
+    let stripped = no_branch
+        .replace("-gguf", "")
+        .replace("unsloth/", "")
+        .replace("meshllm/", "");
+    stripped
+        .chars()
+        .filter(|c| c.is_ascii_alphanumeric())
+        .collect()
+}
+
+/// Backend that calls a local model directly on its skippy HTTP port.
+struct LocalModelBackend {
+    port: u16,
+    http: reqwest::Client,
+}
+
+#[async_trait::async_trait]
+impl moa::ModelBackend for LocalModelBackend {
+    async fn chat_completion(
+        &self,
+        model: &str,
+        messages: &[serde_json::Value],
+        tools: Option<&serde_json::Value>,
+        max_tokens: u32,
+        timeout: std::time::Duration,
+        sampling: moa::SamplingParams,
+    ) -> Result<serde_json::Value, String> {
+        let url = format!("http://127.0.0.1:{}/v1/chat/completions", self.port);
+        let mut body = serde_json::json!({
+            "model": model,
+            "messages": messages,
+            "max_tokens": max_tokens,
+            "temperature": sampling.temperature,
+            "top_p": sampling.top_p,
+            "stream": false,
+            "mesh_hooks": false,
+        });
+        if let Some(tools) = tools {
+            body.as_object_mut()
+                .unwrap()
+                .insert("tools".to_string(), tools.clone());
+        }
+        let resp = self
+            .http
+            .post(&url)
+            .json(&body)
+            .timeout(timeout)
+            .send()
+            .await
+            .map_err(|e| format!("local:{} failed: {e}", self.port))?;
+        let status = resp.status();
+        if !status.is_success() {
+            let text = resp.text().await.unwrap_or_default();
+            return Err(format!("HTTP {status}: {}", &text[..text.len().min(200)]));
+        }
+        resp.json::<serde_json::Value>()
+            .await
+            .map_err(|e| format!("parse: {e}"))
+    }
+}
+
+/// Backend that calls a remote model over the QUIC tunnel.
+struct RemoteModelBackend {
+    node: mesh::Node,
+    peer_id: iroh::EndpointId,
+}
+
+#[async_trait::async_trait]
+impl moa::ModelBackend for RemoteModelBackend {
+    async fn chat_completion(
+        &self,
+        model: &str,
+        messages: &[serde_json::Value],
+        tools: Option<&serde_json::Value>,
+        max_tokens: u32,
+        timeout: std::time::Duration,
+        sampling: moa::SamplingParams,
+    ) -> Result<serde_json::Value, String> {
+        let mut body = serde_json::json!({
+            "model": model,
+            "messages": messages,
+            "max_tokens": max_tokens,
+            "temperature": sampling.temperature,
+            "top_p": sampling.top_p,
+            "stream": false,
+            "mesh_hooks": false,
+        });
+        if let Some(tools) = tools {
+            body.as_object_mut()
+                .unwrap()
+                .insert("tools".to_string(), tools.clone());
+        }
+        let body_bytes = serde_json::to_vec(&body).map_err(|e| format!("serialize: {e}"))?;
+        let http_request = format!(
+            "POST /v1/chat/completions HTTP/1.1\r\n\
+             Host: localhost\r\n\
+             Content-Type: application/json\r\n\
+             Content-Length: {}\r\n\
+             \r\n",
+            body_bytes.len()
+        );
+        let mut raw = http_request.into_bytes();
+        raw.extend_from_slice(&body_bytes);
+
+        let result = tokio::time::timeout(timeout, async {
+            let (mut send, mut recv) = self
+                .node
+                .open_http_tunnel(self.peer_id)
+                .await
+                .map_err(|e| format!("tunnel: {e}"))?;
+            send.write_all(&raw)
+                .await
+                .map_err(|e| format!("send: {e}"))?;
+            send.finish().map_err(|e| format!("finish: {e}"))?;
+            let response = recv
+                .read_to_end(256 * 1024)
+                .await
+                .map_err(|e| format!("recv: {e}"))?;
+            parse_quic_http_response(&response)
+        })
+        .await
+        .map_err(|_| format!("remote timeout after {}s", timeout.as_secs()))?;
+        result
+    }
+}
+
+fn parse_quic_http_response(response: &[u8]) -> Result<serde_json::Value, String> {
+    let s = String::from_utf8_lossy(response);
+    let header_end = s
+        .find("\r\n\r\n")
+        .ok_or_else(|| "malformed HTTP response".to_string())?;
+    let status_line = s[..header_end].lines().next().unwrap_or("");
+    let status: u16 = status_line
+        .split_whitespace()
+        .nth(1)
+        .and_then(|s| s.parse().ok())
+        .unwrap_or(0);
+    if status != 200 {
+        return Err(format!("HTTP {status}: {}", &s[..s.len().min(200)]));
+    }
+    let body = &s[header_end + 4..];
+    serde_json::from_str(body).map_err(|e| format!("parse: {e}"))
+}
+
+/// Send the MoA response as a one-shot SSE stream so SSE-only clients
+/// (like Goose) can consume it. Emits one delta chunk with the full
+/// content, then a `finish_reason: stop` chunk, then `[DONE]`.
+///
+/// `extra_headers` are emitted alongside the standard SSE response headers
+/// (used to attach `x-moa-*` observability headers).
+async fn send_moa_as_sse(
+    mut stream: TcpStream,
+    response: &serde_json::Value,
+    extra_headers: &[(&str, String)],
+) -> std::io::Result<()> {
+    let mut header = String::from(
+        "HTTP/1.1 200 OK\r\nContent-Type: text/event-stream\r\nTransfer-Encoding: chunked\r\nCache-Control: no-cache\r\nConnection: close\r\n",
+    );
+    for (name, value) in extra_headers {
+        header.push_str(name);
+        header.push_str(": ");
+        header.push_str(value);
+        header.push_str("\r\n");
+    }
+    header.push_str("\r\n");
+    stream.write_all(header.as_bytes()).await?;
+
+    let id = response
+        .get("id")
+        .and_then(|v| v.as_str())
+        .unwrap_or("chatcmpl-mesh");
+    let model = response
+        .get("model")
+        .and_then(|v| v.as_str())
+        .unwrap_or(moa::VIRTUAL_MODEL_NAME);
+    let raw_content = response
+        .pointer("/choices/0/message/content")
+        .and_then(|v| v.as_str())
+        .unwrap_or("");
+    let content = strip_think_from_content(raw_content);
+
+    let tool_calls = response
+        .pointer("/choices/0/message/tool_calls")
+        .and_then(|v| v.as_array())
+        .cloned();
+
+    let finish_reason = if tool_calls.is_some() {
+        "tool_calls"
+    } else {
+        "stop"
+    };
+
+    let delta = if let Some(ref tcs) = tool_calls {
+        serde_json::json!({
+            "role": "assistant",
+            "tool_calls": tcs.iter().enumerate().map(|(i, tc)| {
+                serde_json::json!({
+                    "index": i,
+                    "id": tc.get("id").and_then(|v| v.as_str()).unwrap_or("call_0"),
+                    "type": "function",
+                    "function": tc.get("function").cloned().unwrap_or(serde_json::json!({})),
+                })
+            }).collect::<Vec<_>>()
+        })
+    } else {
+        serde_json::json!({ "role": "assistant", "content": content })
+    };
+
+    let chunk = serde_json::json!({
+        "id": id,
+        "object": "chat.completion.chunk",
+        "model": model,
+        "choices": [{
+            "index": 0,
+            "delta": delta,
+            "finish_reason": null,
+        }]
+    });
+    let data = format!("data: {}\n\n", chunk);
+    let framed = format!("{:x}\r\n{}\r\n", data.len(), data);
+    stream.write_all(framed.as_bytes()).await?;
+
+    let stop = serde_json::json!({
+        "id": id,
+        "object": "chat.completion.chunk",
+        "model": model,
+        "choices": [{
+            "index": 0,
+            "delta": {},
+            "finish_reason": finish_reason,
+        }]
+    });
+    let data = format!("data: {}\n\n", stop);
+    let framed = format!("{:x}\r\n{}\r\n", data.len(), data);
+    stream.write_all(framed.as_bytes()).await?;
+
+    let done = "data: [DONE]\n\n";
+    let framed = format!("{:x}\r\n{}\r\n", done.len(), done);
+    stream.write_all(framed.as_bytes()).await?;
+
+    stream.write_all(b"0\r\n\r\n").await?;
+    stream.shutdown().await?;
+    Ok(())
+}
+
+/// Strip `<think>...</think>` tags and orphan `</think>` from content.
+fn strip_think_from_content(text: &str) -> String {
+    let mut result = text.to_string();
+    while let Some(start) = result.find("<think>") {
+        if let Some(end) = result[start..].find("</think>") {
+            result = format!(
+                "{}{}",
+                &result[..start],
+                &result[start + end + "</think>".len()..]
+            );
+        } else {
+            result = result[..start].to_string();
+            break;
+        }
+    }
+    result = result.replace("</think>", "");
+    result.trim().to_string()
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn canonical_base_dedupes_unsloth_and_gguf_variants() {
+        assert_eq!(
+            canonical_base_name("unsloth/Qwen3-8B-GGUF:Q4_K_M"),
+            canonical_base_name("Qwen3-8B-Q4_K_M")
+        );
+        assert_eq!(
+            canonical_base_name("unsloth/Qwen3-8B-GGUF@main:Q4_K_M"),
+            canonical_base_name("Qwen3-8B-Q4_K_M")
+        );
+    }
+
+    #[test]
+    fn canonical_base_keeps_distinct_models_distinct() {
+        assert_ne!(
+            canonical_base_name("unsloth/Qwen3-8B-GGUF:Q4_K_M"),
+            canonical_base_name("unsloth/Qwen3-32B-GGUF:Q4_K_M")
+        );
+        assert_ne!(
+            canonical_base_name("unsloth/Qwen3-32B-GGUF:Q4_K_M"),
+            canonical_base_name("unsloth/MiniMax-M2.5-GGUF:Q4_K_M")
+        );
+    }
+
+    #[test]
+    fn strip_think_handles_simple_block() {
+        assert_eq!(
+            strip_think_from_content("<think>reasoning</think>answer"),
+            "answer"
+        );
+    }
+
+    #[test]
+    fn strip_think_handles_orphan_close_tag() {
+        // Orphan `</think>` is removed but prefix content is preserved.
+        assert_eq!(
+            strip_think_from_content("stuff</think>answer"),
+            "stuffanswer"
+        );
+    }
+
+    #[test]
+    fn strip_think_handles_unclosed_block() {
+        assert_eq!(
+            strip_think_from_content("answer prefix<think>never closed"),
+            "answer prefix"
+        );
+    }
+}
