+++ crates/mesh-llm-host-runtime/src/network/openai/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod ingress;
+pub(crate) mod moa_gateway;
 pub(crate) mod response_adapter;
 pub(crate) mod transport;
